% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortDiagnosticsStudy.R
\docType{package}
\name{cohortDiagnosticsStudy-package}
\alias{cohortDiagnosticsStudy}
\alias{cohortDiagnosticsStudy-package}
\title{cohortDiagnosticsStudy: A Study Package for Generating Cohort Diagnostics}
\description{
A package that generates cohort diagnostics for cohort definitions included in the package.
}
\keyword{internal}
